Add basic SCSS file structure
Add the "style" folder with "base" and "helpers" subfolders to use the 7 in 1 pattern
from SASS guidelines.

The "components" subfolder is not necessary as style for components is self contained.

The "atoms" subfolder is an analog for components that are elements too small to
require a fully fledged vue component.

"Themes" and "Vendors" are not useful for now. "Pages" probably won't
ever be needed as pages are vue components.